refactor: move and rename index to core in src
