+++ package.json
@@ -10,15 +10,15 @@
     "scroll"
   ],
   "main": "dist/scripts.cjs.js",
-  "module": "index.js",
+  "module": "src/core.js",
   "browser": "dist/scripts.js",
   "scripts": {
     "build": "npm-run-all clean scripts",
     "clean": "del dist",
     "scripts": "rollup -c",
     "watch": "nodemon -e js -x 'npm run scripts' -i dist",
-    "lint": "eslint index.js",
-    "lint:fix": "eslint index.js --fix",
+    "lint": "eslint src/*",
+    "lint:fix": "eslint src/* --fix",
     "test": "jest",
     "test:coverage": "jest --coverage",
     "test:watch": "jest --watch",
@@ -52,7 +52,7 @@
     "eslint-plugin-jest": "^23.13.2",
     "jest": "^26.0.1",
     "jest-puppeteer": "^4.4.0",
-    "lodash": "^4.17.19",
+    "lodash.throttle": "^4.1.1",
     "nodemon": "^2.0.4",
     "npm-run-all": "^4.1.5",
     "puppeteer": "^5.0.0",
